Fix indentation for nested lambdas, which are compiled outside the function enclosing them.
If a function is being delayed from adding to , it will decrease the indent of all its children. This aligns its indentation with enclosing function itself, since now they'll be on the same level. Meanwhile, the lambda's compile_return re-adds the indent since the return statement is still part of the function.