Implement closures by adding the following changes:
Lambda definitions result in an array containing the function name and the id of its closure at the time of creation. The lambda function itself takes as its last parameter the closure_id it should use to look up lexically bound variables. Numerous changes have been made to Lexical. Namely, static functions have been added to it to use the right management of what closure_id to use when lexing or binding.