Add compile-with-flag and compile-dir functions to phake library to aid in using it as a build tool.
compile-with-flag toggles a flag for the duration of compiling a single file. However, it does not try to revert the flag back to its original state but instead simply sets it back to False. This may change later. Since there are only two flags as of right now, no-import-lang and executable, I don't think this is a big issue.

compile-dir recursively compiles all Pharen files (.phn) in a directory.

Also, several support functions used by the above are in place.