Add support for default parameter values.
Default parameters are added by wrapping the parameter-value pair in a list. Eg: (arg1 [arg2 bar]). FuncDefNode turns this into a 2-member array and passes that around. Whenever a list of params is being iterated through, the code check if it's an array (meaning a default value is used). The second member of the array is the node of the default value. When building the parameter string, it turns this 2-member array into =value.