Add a  static variable to keep track of the level of ghosting.
Sometimes, ghosting can be turned on/off in the same run. In these cases only the outermost level can turn ghosting off. To solve this, ghost_num keeps track of how many times something is downghost-ed, when it hits 0 it assumes the outermost level is reached.

This will help solve the problem of a macro being run if it's called in the return body of another macro