Add support for object instantiation and method calls on object.
To create an instance of an object the 'new' command will be used: (new MyClass arg1 arg2)

To call a method on an object the '->' command will be used: (-> myobj myfunc arg1 arg2)

Both of those will compile to the normal PHP that would otherwise be handwritten. A special form is needed for method calls so that the -> in PHP can be used instead of call_user_func_array, which hurts performance.