Add AccessModifierNode to allow for public/private/etc. modifiers.
It basically takes a LeafNode and prefixes some body code with it. The body code should usually be a single expression, though because it is a SpecialForm it could support multiple expressions.

It needs to use SpecialForm to use its prefix capabilities. That way the modifier is applied before the indentation.