Add LocalNode.
This simply creates a variable and assigns it a value, without any messing with lexical scoping.

This will allow users to avoid unnecessary lexical baggage when one isn't needed and they just want to create a variable. It can also be used in combination with access modifiers to create fields in classes.