Stops utilizing parsed macro arguments and quoted values inside macros.
It used to be that code that was quoted and passed to a macro as an argument was pre-parsed, so that macros dealt with node trees. This was a problem since macros should be allowed to arbritrarily manipulate code any way it wants to and return code than is parsed as expected.

For example, if you return '(~a cond foo bar) and 'a' happens to contain 'if', Pharen should recognize that this should become an if form. Previously it didn't since the quoted list was already parsed as a regular function call. Now, however, Pharen only tokenizes it, and reparses once it knows the final result of what the quoted code will be like
.
Closes GH-19.