Prevent adding of tmp values when compiling expressions inside lambdas.
Lambdas can have statement forms inside of them, however since lambdas themselves are usually used as expressions anything inside them should be able to deal with tmp values like a regular statement does. Node:: and the like are supposed to be prepended to the statement, but the actual statement is whatever statement the lambda itself is embedded in.

Closes GH-24.